model.ts needed separating
