+++ frontend/src/store/recommendation_raw.ts
@@ -232,60 +232,3 @@ export function getRecommendationCostPerWeek(
 export function getRecommendationType(recommendation: RecommendationRaw) {
   return recommendation.recommenderSubtype;
 }
-
-export const internalStatusMap: Record<string, string> = {
-  ACTIVE: "Applicable",
-  CLAIMED: "In progress",
-  SUCCEEDED: "Success",
-  FAILED: "Failed",
-  DISMISSED: "Dismissed"
-};
-
-export function throwIfInvalidStatus(statusName: string): void {
-  if (!(statusName in internalStatusMap))
-    throw `invalid status name passed: ${statusName}`;
-}
-
-// We don't want to display API status names directly
-//  For details on how this is stored, see RecommendationExtra definition above
-export function getInternalStatusMapping(statusName: string): string {
-  throwIfInvalidStatus(statusName);
-  return internalStatusMap[statusName];
-}
-
-// All data maintained for each recommendation
-export class RecommendationExtra implements RecommendationRaw {
-  // These should not be modified (including inner fields) outside of tests:
-  readonly name: string;
-  readonly description: string;
-  readonly recommenderSubtype: string;
-  readonly primaryImpact: Impact;
-  readonly content: RecommendationContent;
-  readonly stateInfo: RecommendationStateInfo; // original status
-
-  // need to remember them so that v-data-table knows what to sort by
-  readonly costCol: number;
-  readonly projectCol: string;
-  readonly resourceCol: string;
-  readonly typeCol: string;
-
-  // These can be modified:
-  statusCol: string; // follows the current recommendation status, unlike stateInfo
-  errorHeader?: string; // if apply fails,
-  errorDescription?: string; // error details are stored here
-
-  constructor(rec: RecommendationRaw) {
-    this.name = rec.name;
-    this.description = rec.description;
-    this.recommenderSubtype = rec.recommenderSubtype;
-    this.primaryImpact = rec.primaryImpact;
-    this.content = rec.content;
-    this.stateInfo = rec.stateInfo;
-
-    this.costCol = getRecommendationCostPerWeek(rec);
-    this.projectCol = getRecommendationProject(rec);
-    this.resourceCol = getRecommendationResourceShortName(rec);
-    this.typeCol = getRecommendationType(rec);
-    this.statusCol = getInternalStatusMapping(rec.stateInfo.state);
-  }
-}
