+++ frontend/tests/unit/apply.spec.ts
@@ -13,7 +13,8 @@ See the License for the specific language governing permissions and
 limitations under the License. */
 
 import { enableFetchMocks } from "jest-fetch-mock";
-import { RecommendationExtra, getInternalStatusMapping } from "@/store/model";
+import { RecommendationExtra } from "@/store/recommendation_extra";
+import { getInternalStatusMapping } from "@/store/utils";
 import { freshSampleRawRecommendation } from "./sample_recommendation";
 import { rootStoreFactory } from "@/store/root";
 import { recommendationStoreFactory } from "@/store/recommendations";
