+++ frontend/tests/unit/sample_recommendation.ts
@@ -12,7 +12,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License. */
 
-import { RecommendationRaw } from "@/store/model";
+import { RecommendationRaw } from "@/store/recommendation_raw";
 
 const sampleRawRecommendation: RecommendationRaw = {
   content: {
